move ItemUpdater trait into its own file
