+++ items/src/main/scala/weco/api/items/services/ItemUpdater.scala
@@ -1,0 +1,13 @@
+package weco.api.items.services
+
+import weco.catalogue.internal_model.identifiers.{IdState, IdentifierType}
+import weco.catalogue.internal_model.work.Item
+
+import scala.concurrent.Future
+
+trait ItemUpdater {
+  val identifierType: IdentifierType
+  def updateItems(
+                   items: Seq[Item[IdState.Identified]]
+                 ): Future[Seq[Item[IdState.Identified]]]
+}
